api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
pydantic[email]==2.11.0
pydantic-settings==2.8.1
python-multipart==0.0.20
